♻️ Refactor: 대략적인 타입 지정 완료
