+++ src/pages/FolderPage/FolderPage.tsx
@@ -54,6 +54,7 @@ export default function FolderPage({ userId = 1 }) {
     if (modalData) {
       setModalData(modalData);
     }
+    console.log(modalData);
     setIsModalOpened(!isModalOpened);
   };
